aoc_16
This one stumped me for a bit. It took a little while to realize that if the
input as given has the message index as more than half way through the input,
you only really need the numbers after that in this algorithm. That's why
we just have the loop count backwards to build only those number we actually
need.