aoc_18 part 2:
Whew, this was a nice one. After thinking about it for a bit,
it seemed more and more easier. What's more, the way that
the map gets cut into 4s is such a boon to an algorithm like this
where solving a smaller problem 4 times is so much better than
solving a problem that is 4 times as big. Cool stuff